feat: JWT 로그인 시스템 도입, 환자 연동, 불필요 기능 정리
- JWT 로그인/로그아웃 구현 (앱 시작 시 인증 필수, PIN 제거)
- 환자 초대코드 연동 및 백엔드 포탈 API 추가
- 환자 상세 페이지 및 환자 목록 서버 동기화
- PoseDetector: worldLandmarks 편향 수정 (normalized landmarks 통일)
- 설정에서 프로필 이름 변경 기능 추가
- 불필요 기능 제거: 질환검색, 운동 라이브러리, 주변 재활병원, 네이버 API
- 3D 해부 뷰어 → 3D 뷰어 이름 변경

+++ backend/app/models/audit.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import BigInteger, String, Text, DateTime, JSON, func
+from sqlalchemy import Integer, String, Text, DateTime, JSON, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from ..database import Base
@@ -9,7 +9,7 @@ from ..database import Base
 class AuditLog(Base):
     __tablename__ = "audit_logs"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     user_id: Mapped[str | None] = mapped_column(String(36), nullable=True)
     action: Mapped[str] = mapped_column(String(50), nullable=False, index=True)
     resource: Mapped[str] = mapped_column(String(50), nullable=False)
